image_generation/requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
scipy
pillow
huggingface_hub
datasets